Cleanup
git-svn-id: file:///home/asfernandes/fb/repository/sf/svn/firebird/trunk@12015 de594faa-8d1b-4a0c-9a6a-a7de5f8bf859

+++ src/jrd/dmp.cpp
@@ -574,7 +574,7 @@ static void dmp_data(const data_page* page)
 					   i, index->dpg_offset, index->dpg_length,
 					   expanded_length);
 			ib_fprintf(dbg_file, "trans: %d, format: %d, flags: %#x\n\t",
-					   header->rhd_transaction, header->rhd_format,
+					   header->rhd_transaction, (int) header->rhd_format,
 					   header->rhd_flags);
 			if (header->rhd_b_page)
 			{
